separate content into addons (#111)
* update welcome window to include ability to add and remove addons
* add a background image behind the image in combat text
* combat log and combat text from pets is now forwarded to masters
* fixed a configuration issue where arrows were showing during spellcasting if you had a bow equipped
* fixed various null reference issues
* fixed camera jitter when right mouse is down or character is turning
*  combat text behind the camera will be despawned
*  hide nameplates of interactables that are not visible
* prevent unity editor input dumping causing mouse to flip to max zoom on refocus
* fixed a bug that was causing mounts to get agro
* welcome window will now properly load the first time the project is imported and loaded
* Updated the way equipment models can be applied to equipment to make it more modular.  Now UMA support can be added to a game without updating any of the game scriptable objects.

+++ Assets/AnyRPG/Core/System/Scripts/Interactables/StorageContainerComponent.cs
@@ -332,9 +332,11 @@ namespace AnyRPG {
 
         public void ClearAllItems() {
             foreach (InventorySlot inventorySlot in inventorySlots) {
-                // remove the items from the database
-                foreach (InstantiatedItem instantiatedItem in inventorySlot.InstantiatedItems.Values) {
-                    serverDataService.DeleteItemInstance(instantiatedItem);
+                // remove the items from the database in network mode, otherwise they will be orphaned and never cleaned up
+                if (networkManagerServer.ServerModeActive == true) {
+                    foreach (InstantiatedItem instantiatedItem in inventorySlot.InstantiatedItems.Values) {
+                        serverDataService.DeleteItemInstance(instantiatedItem);
+                    }
                 }
                 inventorySlot.RemoveAllItems();
             }
